solutions/java/src/test/java/playground/algorithm/WordSearchTest.java
package playground.algorithm;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class WordSearchTest {

    private WordSearch solution = new WordSearch();

    @Test
    void case_1() {
        assertThat(solution.exist(new char[][]{
                {'A', 'B', 'C', 'E'},
                {'S', 'F', 'C', 'S'},
                {'A', 'D', 'E', 'E'}}, "ABCCED")).isTrue();
    }

    @Test
    void case_2() {
        assertThat(solution.exist(new char[][]{
                {'A', 'B', 'C', 'E'},
                {'S', 'F', 'C', 'S'},
                {'A', 'D', 'E', 'E'}}, "SEE")).isTrue();
    }

    @Test
    void case_3() {
        assertThat(solution.exist(new char[][]{
                {'A', 'B', 'C', 'E'},
                {'S', 'F', 'C', 'S'},
                {'A', 'D', 'E', 'E'}}, "ABCB")).isFalse();
    }
}